fix circular imports in utils
Summary:

As titled

Test Plan:

```
with-proxy test/test_everything.sh
```

Reviewers:

Subscribers:

Tasks:

Tags:
